move runNestedCodeCells to utils
